requirements.txt
requests
pydantic
groq
instructor
python-dotenv

